pyproject.toml
[tool.poetry]
name = "raytracer"
version = "0.1.0"
description = "A simple rendering engine using raytracing"
authors = ["Tim Keller"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
scipy = "^1.11.3"
Pillow = "^10.1.0"
Cython = "^3.0.4"
setuptools = "^68.2.2"
tqdm = "^4.66.1"
numba = "^0.58.1"
snakeviz = "^2.2.0"
numpy-stl = "^3.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
